Move policy::policy into policy
